package restructure + a command byte 'library'
